% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol}
\alias{alcohol}
\title{Per capita alcohol consumption by country and year}
\format{
A tibble with 411 rows and 3 columns:
\describe{
\item{country}{Factor. Country name. Levels are in alphabetical order.}
\item{year}{Integer. Year of measurement (1985 to 2008).}
\item{alcohol}{Numeric. Recorded per capita consumption in litres of pure
alcohol per year.}
}
}
\source{
World Health Organization Global Information System on Alcohol and
Health (GISAH).
\url{https://www.who.int/data/gho/indicator-metadata-registry/imr-details/462}
}
\usage{
alcohol
}
\description{
Recorded per capita alcohol consumption (litres of pure alcohol, adults aged
15 and over) for 190 countries over the period 1985 to 2008.
Sourced from the WHO Global Information System on Alcohol and Health (GISAH).
}
\keyword{datasets}
