% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rats}
\alias{rats}
\title{Rat tumour incidence data}
\format{
A tibble with 71 rows and 3 columns:
\describe{
\item{batch}{Integer. Batch (group) number, 1 to 71.}
\item{m}{Integer. Number of rats in the batch that developed a tumour.}
\item{n}{Integer. Total number of rats in the batch.}
}
}
\source{
Tarone, R.E. (1982). The use of historical control information in
testing for a trend in proportions. \emph{Biometrics}, \strong{38},
215--220.
Reproduced as Table 5.1 in: Gelman, A., Carlin, J.B., Stern, H.S., Dunson,
D.B., Vehtari, A., and Rubin, D.B. (2013). \emph{Bayesian Data Analysis},
3rd ed. Chapman & Hall/CRC.
Raw data at \url{https://sites.stat.columbia.edu/gelman/book/data/rats.asc}.
}
\usage{
rats
}
\description{
Tumour incidence in 71 groups (batches) of female lab rats from a study on
the carcinogenic effects of phenformin.
The data are widely used as a worked example for hierarchical binomial models
and Bayesian partial pooling.
}
\keyword{datasets}
